+++ SmartGreenhouse.WebAPI/Models/ConditionsReading.cs
@@ -9,11 +9,11 @@ namespace SmartGreenhouse.WebAPI.Models
     {
         public int Id { get; set; }
 
-        public int AirTemperature { get; set; }
+        public double AirTemperature { get; set; }
 
-        public int AirHumidity { get; set; }
+        public double AirHumidity { get; set; }
 
-        public int SoilMoisture { get; set; }
+        public double SoilMoisture { get; set; }
 
         public int LightIntensity { get; set; }
